prompts/requirements.txt
browser-use[video]
python-dotenv
google-generativeai
tqdm